Update dx.aspnet.data.js
Added byKeyURL. Enables the user to specify the URL to use when fetching using the /api/controller/id or similar pattern. E.g. using a store productStore and specifying byKeyURL as http://example.com/api/product, a call to productStore.byKey(5) will yield a request;

http://example.com/api/product/5

This is instead of the devexpress default by which it would use the loadURL and append a filter.